refactor: 로그아웃 AuthController에서 UserController로 옮김
