+++ src/main/java/com/study/petory/domain/user/controller/UserController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,6 +21,7 @@ import com.study.petory.domain.user.dto.TokenResponseDto;
 import com.study.petory.domain.user.dto.UpdateUserRequestDto;
 import com.study.petory.domain.user.dto.UserProfileResponseDto;
 import com.study.petory.domain.user.service.UserService;
+import com.study.petory.domain.user.service.UserServiceImpl;
 
 import lombok.RequiredArgsConstructor;
 
@@ -29,6 +31,7 @@ import lombok.RequiredArgsConstructor;
 public class UserController {
 
 	private final UserService userService;
+	private final UserServiceImpl userServiceImpl;
 
 	/**
 	 * [TEST - 로그인]
@@ -82,6 +85,24 @@ public class UserController {
 		userService.updateProfile(currentUser.getEmail(), updateDto);
 
 		return CommonResponse.of(SuccessCode.UPDATED);
+	}
+
+	/**
+	 * [로그아웃 처리]
+	 * AccessToken 을 블랙리스트에 등록하고,
+	 * Redis 에 저장된 RefreshToken 삭제
+	 *
+	 * @param accessToken : "Bearer {accessToken}" 형식의 헤더
+	 * @return 로그아웃 성공 메시지
+	 */
+	@DeleteMapping("/logout")
+	@PreAuthorize("hasAnyRole('USER', 'ADMIN')")
+	public ResponseEntity<CommonResponse<Object>> logout(
+		@RequestHeader("Authorization") String accessToken) {
+
+		userServiceImpl.logout(accessToken);
+
+		return CommonResponse.of(SuccessCode.USER_LOGOUT);
 	}
 
 	/**
